# optimisation of fdk python setup