--- a/Python/tigre/utilities/filtering.py
+++ b/Python/tigre/utilities/filtering.py
@@ -1,7 +1,9 @@
 from __future__ import print_function
 from __future__ import division
+from numpy.core.arrayprint import dtype_is_implied
 from tigre.utilities.parkerweight import parkerweight
 import numpy as np
+from scipy.fftpack  import fft, ifft
 
 import warnings
 
@@ -15,18 +17,34 @@ def filtering(proj,geo,angles,parker,verbose=False):
 
     d=1
     filt=filter(geo.filter,ramp_kernel[0],filt_len,d,verbose=verbose)
-    filt=np.kron(np.ones((np.int64(geo.nDetector[0]),1)),filt)
-    for i in range(angles.shape[0]):
-        fproj=np.zeros((geo.nDetector[0],filt_len),dtype=np.float32)
-        fproj[:,int(filt_len/2-geo.nDetector[1]/2):int(filt_len/2+geo.nDetector[1]/2)]=proj[i]
-        fproj=np.fft.fft(fproj,axis=1)
+    filt=np.kron(np.ones((np.int64(geo.nDetector[0]),1),dtype=np.float32),filt)
 
+    padding = int((filt_len-geo.nDetector[1])//2 )
+    scale_factor = (geo.DSD[0]/geo.DSO[0]) * (2 * np.pi/ len(angles)) / ( 4 * geo.dDetector[0] ) 
+
+    #filter 2 projection at a time packing in to complex container
+    fproj=np.empty((geo.nDetector[0],filt_len),dtype=np.complex64)
+    for i in range(0,angles.shape[0]-1,2):
+        fproj.fill(0)
+        fproj.real[:,padding:padding+geo.nDetector[1]]=proj[i]
+        fproj.imag[:,padding:padding+geo.nDetector[1]]=proj[i+1]
+
+        fproj=fft(fproj,axis=1)
         fproj=fproj*filt
+        fproj=ifft(fproj,axis=1)
 
-        fproj=np.real(np.fft.ifft(fproj,axis=1))
-        proj[i]=fproj[:,int(filt_len/2-geo.nDetector[1]/2):int(filt_len/2+geo.nDetector[1]/2)]/2/geo.dDetector[0]*(2*np.pi/
-                                                                                                 len(angles)
-                                                                                                 )/2*(geo.DSD[0]/geo.DSO[0])
+        proj[i]=fproj.real[:,padding:padding+geo.nDetector[1]] * scale_factor
+        proj[i+1]=fproj.imag[:,padding:padding+geo.nDetector[1]] * scale_factor
+
+    #if odd number of projections filter last solo
+    if angles.shape[0] % 2:
+        fproj.fill(0)
+        fproj.real[:,padding:padding+geo.nDetector[1]]=proj[angles.shape[0]-1]
+
+        fproj=fft(fproj,axis=1)
+        fproj=fproj*filt
+        fproj=np.real(ifft(fproj,axis=1))     
+        proj[angles.shape[0]-1]=fproj[:,padding:padding+geo.nDetector[1]] * scale_factor
 
     return proj
 
@@ -43,29 +61,29 @@ def filter(filter,kernel,order,d,verbose=False):
 
     filt=f_kernel[:int((order/2)+1)]
     w=2*np.pi*np.arange(len(filt))/order
-    if filter not in ['ram_lak','shepp_logan','cosine','hamming','hann',None]:
-        raise ValueError('filter not recognised: '+str(filter))
 
     if filter in {'ram_lak', None}:
         if filter is None:
             if verbose:
                 warnings.warn('no filter selected, using default ram_lak')
         pass
-    if filter=='shepp_logan':
+    elif filter=='shepp_logan':
         filt[1:]*=(np.sin(
                                 w[1:]/(2*d))/(w[1:]/(2*d)
                                               )
                            )
-    if filter=='cosine':
+    elif filter=='cosine':
         filt[1:]*=np.cos(w[1:]/(2*d))
-    if filter=='hamming':
+    elif filter=='hamming':
         filt[1:]*=(.54+.46*np.cos(w[1:]/d))
-    if filter =='hann':
+    elif filter =='hann':
         filt[1:]*=(1+np.cos(w[1:])/d)/2
+    else:
+        raise ValueError('filter not recognised: '+str(filter))
 
     filt[w>np.pi*d]=0
     filt=np.hstack((filt,filt[1:-1][::-1]))
-    return filt
+    return filt.astype(np.float32)
 
 
 def nextpow2(n):