@gfardell 2021-09-17
> There are some conflicts with the current master (I assume yours is based on TIGRE 2.0 or something from January). Mind trying to solve those? Happy to help if you have any doubts.

Yep, just doing it now. Of course I was in 2.1 :D

I had a go with IPP and OpenMP and got the filtering time down to <2s for that size dataset. So I think we'll likely implement that in CIL and just call you backprojectors. But I thought you might as well benefit from the simple changes otherwise!