@AnderBiguri 2021-09-17
Absolutely, let me double check if it runs well on a random dataset that I have around before merging! Thanks!